Rename 2.js in 3.js
