--- sketch/0/2.js
@@ -1,139 +1,0 @@
-var noise = new SimplexNoise();
-var vizInit = function (){
-  
-  //var audio = document.getElementById("audio");
-
-  audio.src = '../audio/Iosonouncane_Hajar.mp3';
-  audio.play();
-  play();
-  
-  
-function play() {
-    var context = new AudioContext();
-    var src = context.createMediaElementSource(audio);
-    var analyser = context.createAnalyser();
-    src.connect(analyser);
-    analyser.connect(context.destination);
-    analyser.fftSize = 512;
-    var bufferLength = analyser.frequencyBinCount;
-    var dataArray = new Uint8Array(bufferLength);
-    var scene = new THREE.Scene();
-    var group = new THREE.Group();
-    var camera = new THREE.PerspectiveCamera(45, window.innerWidth / window.innerHeight, 0.1, 1000);
-    camera.position.set(0,0,300);
-    camera.lookAt(scene.position);
-    scene.add(camera);
-
-    var renderer = new THREE.WebGLRenderer({ alpha: false, antialias: true });
-    renderer.setSize(window.innerWidth, window.innerHeight);
-
-    
-
-    var icosahedronGeometry = new THREE.IcosahedronGeometry(0.1, 4); //(radius, details)
-    var lambertMaterial = new THREE.MeshLambertMaterial({
-        color: 0xffffff,
-        wireframe: false
-    });
-
-    var ball = new THREE.Mesh(icosahedronGeometry, lambertMaterial);
-    ball.position.set(0, 0, 0); //centro di rotazione su se stessa
-    group.add(ball);
-
-    var ambientLight = new THREE.AmbientLight(0xe6e3e3);
-    scene.add(ambientLight);
-
-    var spotLight = new THREE.SpotLight(0xbb192a);
-    spotLight.intensity = 0.5;
-    spotLight.position.set(-10, 40, 10);
-    spotLight.lookAt(ball);
-    spotLight.castShadow = true;
-    scene.add(spotLight);
-    
-    scene.add(group);
-
-    document.getElementById('out').appendChild(renderer.domElement);
-
-    window.addEventListener('resize', onWindowResize, false);
-
-    render();
-
-    function render() {
-      analyser.getByteFrequencyData(dataArray);
-
-      var lowerHalfArray = dataArray.slice(0, (dataArray.length/2) - 1);
-      var upperHalfArray = dataArray.slice((dataArray.length/2) - 1, dataArray.length - 1);
-
-      var overallAvg = avg(dataArray);
-      var lowerMax = max(lowerHalfArray);
-      var lowerAvg = avg(lowerHalfArray);
-      var upperMax = max(upperHalfArray);
-      var upperAvg = avg(upperHalfArray);
-
-      var lowerMaxFr = lowerMax / lowerHalfArray.length;
-      var lowerAvgFr = lowerAvg / lowerHalfArray.length;
-      var upperMaxFr = upperMax / upperHalfArray.length;
-      var upperAvgFr = upperAvg / upperHalfArray.length;
-
-     
-      
-      makeRoughBall(ball, modulate(Math.pow(lowerMaxFr, 0.8), 0, 1, 0, 8), modulate(upperAvgFr, 0, 1, 0, 4));
-
-      group.rotation.y += 0.005;
-      renderer.render(scene, camera);
-      requestAnimationFrame(render);
-    }
-
-    function onWindowResize() {
-        camera.aspect = window.innerWidth / window.innerHeight;
-        camera.updateProjectionMatrix();
-        renderer.setSize(window.innerWidth, window.innerHeight);
-    }
-
-    function makeRoughBall(mesh, bassFr, treFr) {
-        mesh.geometry.vertices.forEach(function (vertex, i) {
-            var offset = mesh.geometry.parameters.radius;
-            var amp = 7;
-            var time = window.performance.now();
-            vertex.normalize();
-            var rf = 0.00001;
-            var distance = (offset + bassFr ) + noise.noise3D(vertex.x + time *rf*7, vertex.y +  time*rf*8, vertex.z + time*rf*9) * amp * treFr;
-            vertex.multiplyScalar(distance);
-        });
-        mesh.geometry.verticesNeedUpdate = true;
-        mesh.geometry.normalsNeedUpdate = true;
-        mesh.geometry.computeVertexNormals();
-        mesh.geometry.computeFaceNormals();
-    }
-
-    
-
-    audio.play();
-  };
-}
-
-window.onload = vizInit();
-
-document.body.addEventListener('touchend', function(ev) { context.resume(); });
-
-
-
-
-
-function fractionate(val, minVal, maxVal) {
-    return (val - minVal)/(maxVal - minVal);
-}
-
-function modulate(val, minVal, maxVal, outMin, outMax) {
-    var fr = fractionate(val, minVal, maxVal);
-    var delta = outMax - outMin;
-    return outMin + (fr * delta);
-}
-
-function avg(arr){
-    var total = arr.reduce(function(sum, b) { return sum + b; });
-    return (total / arr.length);
-}
-
-function max(arr){
-    return arr.reduce(function(a, b){ return Math.max(a, b); })
-}
